Alter the calculateHash function to update a knot
Before the change, this function returned the hash of a given gameboard.
The result, however, could not be assigned to the gameboardHash member
of a knot. To do this, it would have been necessary to copy it char by
char.
In order to avoid this, the function now takes a knot as an argument and
sets its gameboardHash according to its gameboard.

Also the function was not called in the right place; the gameboardHash
member would have been used in the checkForDuplicate function before it
was set.